# internal/course/lesson_handler.go
package course

import (
	"net/http"

	"github.com/go-chi/chi/v5"

	"github.com/nguyensongtai/lingora-api/internal/api"
	"github.com/nguyensongtai/lingora-api/internal/httpx"
	"github.com/nguyensongtai/lingora-api/internal/vocabulary"
)

func (h *Handler) createLesson(w http.ResponseWriter, r *http.Request) {
	var body api.LessonCreate
	if err := httpx.DecodeJSON(w, r, &body); err != nil {
		writeMalformed(w, err)
		return
	}

	courseID := chi.URLParam(r, "courseID")
	created, err := h.service.CreateLesson(r.Context(), courseID, LessonCreateParams{
		Slug:  body.Slug,
		Title: body.Title,
	})
	if err != nil {
		writeError(w, r, err)
		return
	}

	w.Header().Set("Location", "/v1/lessons/"+created.ID)
	httpx.JSON(w, http.StatusCreated, toAPILesson(created))
}

func (h *Handler) getLesson(w http.ResponseWriter, r *http.Request) {
	found, err := h.service.GetLessonDetail(r.Context(), ViewerFrom(r), chi.URLParam(r, "lessonID"))
	if err != nil {
		writeError(w, r, err)
		return
	}

	lesson := toAPILesson(found.Lesson)
	httpx.JSON(w, http.StatusOK, api.LessonDetail{
		Id:         lesson.Id,
		CourseId:   lesson.CourseId,
		Slug:       lesson.Slug,
		Title:      lesson.Title,
		Summary:    lesson.Summary,
		Position:   lesson.Position,
		CreatedAt:  lesson.CreatedAt,
		UpdatedAt:  lesson.UpdatedAt,
		Blocks:     toAPIBlocks(found.Blocks),
		Vocabulary: toAPIWords(found.Words),
	})
}

func (h *Handler) replaceLessonBlocks(w http.ResponseWriter, r *http.Request) {
	var body api.LessonBlockList
	if err := httpx.DecodeJSON(w, r, &body); err != nil {
		httpx.Error(w, http.StatusBadRequest, httpx.CodeMalformed, "Body không phải JSON hợp lệ.", nil)
		return
	}

	blocks := make([]Block, 0, len(body.Blocks))
	for _, item := range body.Blocks {
		blocks = append(blocks, Block{
			Kind:    BlockKind(item.Kind),
			Body:    item.Body,
			TextEN:  item.TextEn,
			TextVI:  item.TextVi,
			Speaker: item.Speaker,
		})
	}

	if err := h.service.ReplaceBlocks(r.Context(), chi.URLParam(r, "lessonID"), blocks); err != nil {
		writeError(w, r, err)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

// toAPIBlocks giữ slice rỗng thành [] chứ không null: spec khai báo mảng và
// phía trước lặp thẳng trên nó.
// toAPIWords luôn trả mảng, kể cả rỗng: spec khai báo mảng bắt buộc và phía
// trước lặp thẳng trên nó.
func toAPIWords(words []vocabulary.Entry) []api.VocabularyEntry {
	items := make([]api.VocabularyEntry, 0, len(words))
	for _, word := range words {
		items = append(items, vocabulary.ToAPIEntry(word))
	}
	return items
}

func toAPIBlocks(blocks []Block) []api.LessonBlock {
	out := make([]api.LessonBlock, 0, len(blocks))
	for _, block := range blocks {
		id := block.ID
		out = append(out, api.LessonBlock{
			Id:      &id,
			Kind:    api.LessonBlockKind(block.Kind),
			Body:    block.Body,
			TextEn:  block.TextEN,
			TextVi:  block.TextVI,
			Speaker: block.Speaker,
		})
	}
	return out
}

func (h *Handler) updateLesson(w http.ResponseWriter, r *http.Request) {
	var body api.LessonUpdate
	if err := httpx.DecodeJSON(w, r, &body); err != nil {
		writeMalformed(w, err)
		return
	}

	updated, err := h.service.UpdateLesson(r.Context(), chi.URLParam(r, "lessonID"), LessonUpdateParams{
		Slug:  body.Slug,
		Title: body.Title,
	})
	if err != nil {
		writeError(w, r, err)
		return
	}
	httpx.JSON(w, http.StatusOK, toAPILesson(updated))
}

func (h *Handler) deleteLesson(w http.ResponseWriter, r *http.Request) {
	if err := h.service.DeleteLesson(r.Context(), chi.URLParam(r, "lessonID")); err != nil {
		writeError(w, r, err)
		return
	}
	httpx.NoContent(w)
}

func (h *Handler) reorderLessons(w http.ResponseWriter, r *http.Request) {
	var body api.LessonOrder
	if err := httpx.DecodeJSON(w, r, &body); err != nil {
		writeMalformed(w, err)
		return
	}

	if err := h.service.ReorderLessons(r.Context(), chi.URLParam(r, "courseID"), body.LessonIds); err != nil {
		writeError(w, r, err)
		return
	}
	httpx.NoContent(w)
}

func toAPILesson(item Lesson) api.Lesson {
	return api.Lesson{
		Id:        item.ID,
		CourseId:  item.CourseID,
		Slug:      item.Slug,
		Title:     item.Title,
		Summary:   item.Summary,
		Position:  item.Position,
		CreatedAt: item.CreatedAt,
		UpdatedAt: item.UpdatedAt,
	}
}