# openapi.yaml
openapi: 3.0.3
info:
  title: Lingora API
  version: 0.1.0
  description: |
    API nội dung học tiếng Anh của Lingora.

    Route đọc là công khai. Route ghi cần access token JWT HS256 có claim
    `role: admin`.
servers:
  - url: http://localhost:8080/v1
    description: Máy dev
  - url: https://api.lingora.vn/v1
    description: Production
...
paths:
...
  /courses/{courseId}/lessons:
    parameters:
      - $ref: '#/components/parameters/CourseId'

    get:
      tags: [courses]
      summary: Liệt kê bài học của một khoá
      description: |
        Sắp xếp theo `position` tăng dần. Khoá không tồn tại trả về 404; khoá có
        thật nhưng chưa có bài trả về mảng rỗng.

        Bài không có trạng thái riêng, nó thừa hưởng của khoá: bài thuộc khoá
        `draft` chỉ admin đọc được, người khác cũng nhận 404.
      operationId: listCourseLessons
      responses:
        '200':
          description: Danh sách bài học
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [courses]
      summary: Thêm bài học vào khoá
      description: Bài mới luôn đứng cuối; đổi thứ tự bằng endpoint order.
      operationId: createLesson
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonCreate'
      responses:
        '201':
          description: Bài học đã được tạo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /courses/{courseId}/lessons/order:
    parameters:
      - $ref: '#/components/parameters/CourseId'

    put:
      tags: [courses]
      summary: Sắp xếp lại bài học trong khoá
      description: |
        Danh sách phải liệt kê đúng tập bài hiện có của khoá — thiếu, thừa hay
        trùng đều bị từ chối. Vị trí mới lấy từ chính thứ tự của mảng.
      operationId: reorderLessons
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonOrder'
      responses:
        '204':
          description: Đã sắp xếp lại
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /lessons/{lessonId}:
    parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          x-go-type: string

    get:
      tags: [courses]
      summary: Lấy bài học theo id
      description: |
        Bài không có trạng thái riêng, nó thừa hưởng của khoá: bài thuộc khoá
        `draft` chỉ admin đọc được, người khác nhận 404.
      operationId: getLesson
      responses:
        '200':
          description: Bài học kèm nội dung
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [courses]
      summary: Cập nhật một phần bài học
      description: |
        position không sửa được ở đây; dùng endpoint order của khoá.
      operationId: updateLesson
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonUpdate'
      responses:
        '200':
          description: Bài học sau khi cập nhật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags: [courses]
      summary: Xoá mềm bài học
      operationId: deleteLesson
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Đã xoá
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /lessons/{lessonId}/blocks:
    parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          x-go-type: string

    put:
      tags: [courses]
      summary: Thay toàn bộ nội dung của một bài
      description: |
        Nhận cả danh sách chứ không sửa từng khối: soạn bài là việc kéo thả và
        viết lại, nên gửi trọn trạng thái mong muốn đơn giản hơn nhiều so với
        một chuỗi thao tác thêm/sửa/xoá/đổi chỗ.

        `position` không có trong body — thứ tự chính là thứ tự của mảng.
      operationId: replaceLessonBlocks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonBlockList'
      responses:
        '204':
          description: Đã thay nội dung
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /me/practice/session:
    get:
      tags: [practice]
      summary: Dựng một phiên luyện tập
      operationId: getPracticeSession
      security:
        - bearerAuth: []
      parameters:
        - name: size
          in: query
          required: false
          description: |
            Số câu muốn có, mặc định 10, trần 30. Ngoài khoảng thì kẹp về biên.
            Bỏ qua khi có lesson_id: luyện trong bài luôn hỏi hết từ của bài.
          schema:
            type: integer
        - name: lesson_id
          in: query
          required: false
          description: |
            Luyện riêng từ của một bài, kể cả khi người học chưa bấm "Đã xong".
            Không có thì luyện từ vốn từ đã mở khoá, ưu tiên từ đến hạn ôn.
          schema:
            type: string
            format: uuid
            x-go-type: string
      responses:
        '200':
          description: Phiên luyện tập
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /me/practice/answers:
    post:
      tags: [practice]
      summary: Chấm một câu trả lời
      operationId: checkPracticeAnswer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PracticeAnswer'
      responses:
        '200':
          description: Kết quả chấm
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token HS256 với claim `role: admin`."
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
        x-go-type: string
...
  schemas:
...
    VocabularyEntry:
      type: object
      required: [id, lesson_id, word, ipa, meaning, example, example_vi, position, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          x-go-type: string
        lesson_id:
          type: string
          format: uuid
          x-go-type: string
        word:
          type: string
          maxLength: 100
        ipa:
          type: string
          maxLength: 100
        meaning:
          type: string
          maxLength: 300
        example:
          type: string
          maxLength: 500
        example_vi:
          type: string
          maxLength: 500
        position:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CourseLevel:
      type: string
      description: Trình độ theo khung CEFR.
      enum: [A1, A2, B1, B2, C1, C2]
...
    Lesson:
      type: object
      required: [id, course_id, slug, title, summary, position, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          x-go-type: string
        course_id:
          type: string
          format: uuid
          x-go-type: string
        slug:
          type: string
        title:
          type: string
        summary:
          type: string
          description: Đoạn dẫn ngắn hiện ngay dưới tiêu đề bài; có thể rỗng.
        position:
          type: integer
          format: int32
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LessonBlockKind:
      type: string
      enum: [note, example, dialogue]
      description: |
        `note` là đoạn giải thích tiếng Việt, `example` là câu mẫu Anh–Việt,
        `dialogue` giống `example` nhưng có tên người nói.
    LessonBlock:
      type: object
      required: [kind, body, text_en, text_vi, speaker]
      description: |
        Mỗi dạng chỉ dùng một phần các trường, phần còn lại phải là chuỗi rỗng:

        | kind | trường bắt buộc | phải rỗng |
        | --- | --- | --- |
        | `note` | `body` | `text_en`, `text_vi`, `speaker` |
        | `example` | `text_en` | `body`, `speaker` |
        | `dialogue` | `text_en`, `speaker` | `body` |

        Database có ràng buộc bắt đúng hình dạng này, nên gửi sai là 400 chứ
        không phải một khối trống giữa bài.
      properties:
        id:
          type: string
          format: uuid
          x-go-type: string
          description: Chỉ có khi đọc; lúc ghi thì bỏ qua vì cả danh sách được thay mới.
        kind:
          $ref: '#/components/schemas/LessonBlockKind'
        body:
          type: string
        text_en:
          type: string
        text_vi:
          type: string
        speaker:
          type: string
    LessonBlockList:
      type: object
      required: [blocks]
      description: |
        Thay TOÀN BỘ nội dung của bài. Gửi mảng rỗng là xoá sạch nội dung — đó
        là cách duy nhất để làm việc đó, và nó có chủ đích.
      properties:
        blocks:
          type: array
          items:
            $ref: '#/components/schemas/LessonBlock'
    LessonDetail:
      allOf:
        - $ref: '#/components/schemas/Lesson'
        - type: object
          required: [blocks, vocabulary]
          properties:
            blocks:
              type: array
              description: Nội dung bài theo đúng thứ tự hiển thị.
              items:
                $ref: '#/components/schemas/LessonBlock'
            vocabulary:
              type: array
              description: |
                Từ vựng bài này dạy, theo thứ tự của người soạn. Trả cho mọi
                người đọc được bài — khác với /me/vocabulary, vốn chỉ gồm từ của
                những bài đã học xong.
              items:
                $ref: '#/components/schemas/VocabularyEntry'
    LessonCreate:
      type: object
      required: [slug, title]
      properties:
        slug:
          type: string
          maxLength: 120
          pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
        title:
          type: string
          maxLength: 200
    LessonUpdate:
      type: object
      minProperties: 1
      properties:
        slug:
          type: string
          maxLength: 120
          pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
        title:
          type: string
          maxLength: 200
        summary:
          type: string
          maxLength: 1000
    LessonOrder:
      type: object
      required: [lesson_ids]
      properties:
        lesson_ids:
          type: array
          minItems: 1
          description: Toàn bộ id bài của khoá, theo thứ tự mong muốn.
          items:
            type: string
            format: uuid
            x-go-type: string
    LessonList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Lesson'
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - validation_error
            - malformed_body
            - unauthorized
            - forbidden
            - not_found
            - method_not_allowed
            - conflict
            - rate_limited
            - internal_error
        message:
          type: string
          description: Câu thông báo tiếng Việt, hiển thị được cho người dùng.
        details:
          type: object
          additionalProperties:
            type: string
          description: Lỗi theo từng field, chỉ có với validation_error.
...
    PracticeKind:
      type: string
      enum: [multiple_choice, fill_blank, listen_choose]
    PracticeQuestion:
      type: object
      required: [entry_id, kind, level, word, prompt, hint, options]
      properties:
        entry_id:
          type: string
          format: uuid
          x-go-type: string
        kind:
          $ref: '#/components/schemas/PracticeKind'
        level:
          $ref: '#/components/schemas/CourseLevel'
        word:
          type: string
          description: |
            Từ tiếng Anh. Với `listen_choose`, phía trước cần nó để máy đọc lên
            — nghĩa là đáp án nằm sẵn trong trang. Đó là hệ quả không tránh
            được khi phát âm bằng `speechSynthesis` thay vì file audio.
        prompt:
          type: string
          description: |
            Câu dẫn. `multiple_choice`: chính từ đó. `fill_blank`: câu ví dụ đã
            khoét chỗ trống. `listen_choose`: chuỗi rỗng, câu dẫn là âm thanh.
        hint:
          type: string
          description: Gợi ý thêm; hiện chỉ `fill_blank` dùng, để hiện nghĩa tiếng Việt.
        options:
          type: array
          description: Rỗng với `fill_blank` vì dạng đó gõ tay.
          items:
            type: string
    PracticeSession:
      type: object
      required: [questions]
      description: |
        Câu hỏi sinh tại chỗ từ vốn từ người học đã mở khoá — không có bảng câu
        hỏi nào để soạn. Chưa đủ từ thì `questions` rỗng; đó là trạng thái bình
        thường của màn hình, không phải lỗi.
      properties:
        questions:
          type: array
          items:
            $ref: '#/components/schemas/PracticeQuestion'
    PracticeAnswer:
      type: object
      required: [entry_id, kind, answer]
      properties:
        entry_id:
          type: string
          format: uuid
          x-go-type: string
        kind:
          $ref: '#/components/schemas/PracticeKind'
        answer:
          type: string
          description: Lựa chọn đã chọn, hoặc chuỗi đã gõ. So sánh bỏ qua hoa/thường và khoảng trắng thừa.
        lesson_id:
          type: string
          format: uuid
          x-go-type: string
          description: |
            Có khi câu hỏi đến từ lượt luyện trong bài (session với lesson_id).
            Khi đó câu sai KHÔNG đụng tới lịch ôn: đây là lần gặp đầu, không
            phải ôn.
    PracticeResult:
      type: object
      required: [correct, expected, penalised]
      properties:
        correct:
          type: boolean
        expected:
          type: string
          description: Đáp án đúng, trả về cả khi làm đúng để người học đối chiếu.
        penalised:
          type: boolean
          description: |
            Câu sai đẩy từ về đầu hàng đợi ôn. Trả lời ĐÚNG không kéo dài
            khoảng cách ôn — luyện dồn một buổi không được biến một từ vừa gặp
            thành "thành thạo". Luôn là false với luyện tập trong bài.
...
  responses:
...
    BadRequest:
      description: Dữ liệu vào không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Thiếu token hoặc token không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Token hợp lệ nhưng không đủ quyền
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Không tìm thấy
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Xung đột với dữ liệu đang có
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'